requirements.txt
discord.py[voice]==2.2.2
PyQt5==5.15.11
sounddevice==0.4.6
numpy>=1.20.0
